changed the way stack replacement works
Before, a replace on an empty stack did nothing. Now, it pushes. The benefit is that it can be *certain* that replacing will actually set the top element of the stack, making replacement a more useful operation in that regard. I noticed that the old way replace worked was inconvenient while creating tests for it, since I had to keep computing the probability that the stack was empty.